% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverleaf.R
\name{leaf_output}
\alias{leaf_output}
\title{Server leaflet map}
\usage{
leaf_output(type_choose, date_choose)
}
\arguments{
\item{inputId}{The id will access the output from server}

\item{lable}{The inpt name}

\item{choices}{Three type of the cases}
}
\value{

}
\description{
This function builds leaflet map
}
\examples{
 leaf_output("confirmed", "2020-02-10")


}
