% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{menu}
\alias{menu}
\title{SHiny Dashboard Menu}
\source{
Icon you would like to choose can be found here:
\url{https://fontawesome.com/icons?d=gallery&q=covid-19&m=free}
}
\usage{
menu(id, tab, name)
}
\arguments{
\item{id}{Name of the menu}

\item{name}{The icon name from Frontawsome}

\item{teb}{Tab name of the menu}
}
\description{
Add menu and icon in the sidebar menu
}
\examples{
menu("Plot", "Plot", "chart-bar")

}
