% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyllispkg-package.R
\docType{package}
\name{phyllispkg-package}
\alias{phyllispkg}
\alias{phyllispkg-package}
\title{phyllispkg: My first R package}
\description{
Contains data and my first Shinyapp.
}
\author{
\strong{Maintainer}: Peimin Lin \email{plin0018@student.monash.edu}

}
\keyword{internal}
