% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiselect.R
\name{radio_input}
\alias{radio_input}
\title{Select RadioButton type}
\usage{
radio_input(inputId, label, choices)
}
\arguments{
\item{inputId}{The id will access the output from server}

\item{choices}{Three type of the cases}

\item{lable}{The inpt name}
}
\value{

}
\description{
This function builds a radioButton to select the type of cases
}
