% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\docType{data}
\name{tidy}
\alias{tidy}
\title{Covid-19 total cases in 10 countries}
\format{
a tibble with 27072 observations and 7 variables
\itemize{
\item \strong{type}: type of COVID-19 cases, confirmed, death, recovered
\item \strong{date}: date from 2020-01-22 to 2020-07-31
\item \strong{country}: 10 main countries, US, Australia, China, Italy, Belgium, Spain, Germany, Korean, Japan, England
\item \strong{total_cases}: number of cases
}
}
\source{
The data are provided by the Johns Hopkins University Center for Systems Science and Engineering (JHU CSSE) and Rami Krispin and other 4 contributor provided a tidy format of the data as r package coronavirus.
}
\usage{
tidy
}
\description{
This data set contains total numbers of 10 main countries of difference type cases every day.
}
\keyword{datasets}
