% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiselect.R
\name{type_input}
\alias{type_input}
\title{Select date}
\usage{
type_input(inputid, type)
}
\arguments{
\item{InputId}{The unqique id for user to interface the shiny app}

\item{date}{The date would like to select}
}
\value{

}
\description{
This function builds a date input to select the date
}
\examples{



}
