% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverleaf.R
\name{leaf_output}
\alias{leaf_output}
\title{Server leaflet map}
\usage{
leaf_output(data, type_choose, date_choose)
}
\arguments{
\item{data}{Data set choose for map}

\item{type_choose}{The type choose}

\item{date_choose}{The input date}
}
\value{

}
\description{
This function builds leaflet map
}
