% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{map}
\alias{map}
\title{Covid-19 map}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 27072 rows and 7 columns.
}
\source{
The data are provided by the Johns Hopkins University Center for Systems Science and Engineering (JHU CSSE) and Rami Krispin and other 4 contributor provided a tidy format of the data as r package coronavirus.
}
\usage{
map
}
\description{
This data set contains longitude and latitude of 10 main countries
}
\keyword{datasets}
